Remove index only if it exists

+++ db/main/migrate/20171103000000_update_index_on_abuses.rb
@@ -2,7 +2,7 @@ class UpdateIndexOnAbuses < ActiveRecord::Migration
   self.disable_ddl_transaction!
 
   def up
-    remove_index :abuses, :owner
+    remove_index :abuses, :owner if index_exists?(:abuses, :owner)
     add_index :abuses, [:owner_id, :owner_type, :level], unique: true, algorithm: :concurrently
   end
